Fix logic around misleading log for skipping deletion of unregistered nodes

+++ cluster-autoscaler/cloudprovider/mcm/mcm_cloud_provider.go
@@ -425,6 +425,9 @@ func (ngImpl *nodeGroup) ForceDeleteNodes(nodes []*apiv1.Node) error {
 		}
 		toBeDeletedMachineInfos = append(toBeDeletedMachineInfos, *mInfo)
 	}
+	if len(toBeDeletedMachineInfos) == 0 {
+		return ErrNoNodesToDelete
+	}
 	return ngImpl.deleteMachines(toBeDeletedMachineInfos)
 }
